move phases into internal package
