+++ migration/state/newstate/internal/common/constants.go
