+++ migration/state/newstate/internal/common/counter.go
