+++ migration/state/newstate/internal/common/counter_test.go
