+++ migration/state/newstate/internal/headstate/migrator_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/NethermindEth/juno/core/state"
 	"github.com/NethermindEth/juno/db"
 	"github.com/NethermindEth/juno/db/memory"
-	"github.com/NethermindEth/juno/migration/state/headstate"
+	"github.com/NethermindEth/juno/migration/state/newstate/internal/headstate"
 	"github.com/NethermindEth/juno/utils/log"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
