+++ migration/state/newstate/internal/history/class_hash_ingestor.go
@@ -10,7 +10,7 @@ import (
 	"github.com/NethermindEth/juno/db/dbutils"
 	"github.com/NethermindEth/juno/migration/pipeline"
 	"github.com/NethermindEth/juno/migration/semaphore"
-	"github.com/NethermindEth/juno/migration/state/common"
+	"github.com/NethermindEth/juno/migration/state/newstate/internal/common"
 )
 
 type classHashIngestor struct {
